pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
name = "sfmanager"
version = "0.1.4"
authors = [
	{name="GrandTheBest", email="grandinfo-cm@gmail.com"},
]
description = "This library is not that useful. Perhaps someday it will become great."
readme = "README.md"
requires-python = ">=3.10.0"
dependencies = [
	"Pillow>=9.0.1",
   "loguru>=0.7.2"
]
classifiers = [
   "Programming Language :: Python :: 3",
   "License :: OSI Approved :: MIT License",
   "Operating System :: OS Independent",
]
[project.scripts]
sfmanager = "sfmanager:checkUpdates"
[project.urls]
Homepage = "https://github.com/GrandTheBest/sfmanager"
Issues = "https://github.com/GrandTheBest/sfmanager/issues"